# openapi.json
{
  "components": {
    "schemas": {
...
      "CalendarDate": {
        "description": "A calendar date, `YYYY-MM-DD`. Not an instant: an accounting date carries no time and no timezone, because which fiscal period an entry lands in must not depend on the reader’s.",
        "examples": [
          "2026-03-31"
        ],
        "format": "date",
        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
        "type": "string"
      },
      "CalendarDateInput": {
        "description": "A calendar date, `YYYY-MM-DD`. Not an instant: an accounting date carries no time and no timezone, because which fiscal period an entry lands in must not depend on the reader’s.",
        "examples": [
          "2026-03-31"
        ],
        "format": "date",
        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
        "type": "string"
      },
...
      "CreateOrgRequestInput": {
        "additionalProperties": false,
        "description": "Creates an organization with the calling user as its Owner. No slug — it is derived from the name.",
        "properties": {
          "fiscalYearStartMonth": {
            "description": "The calendar month the org’s fiscal year begins in, 1–12. April, July, and October are all common. Defaults to January where the field is optional.",
            "maximum": 12,
            "minimum": 1,
            "type": "integer"
          },
          "name": {
            "description": "Length is bounded by the server (`orgs.name` is `VARCHAR(255)`) and reported as a `validation_failed` naming `name`.",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "The body of every non-2xx response.",
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "description": "Stable machine-readable failure code. Integrators branch on this, not on the HTTP status: `conflict` and `idempotency_key_conflict` are both 409 and have different recoveries. Codes are never renamed and never change status.",
                "enum": [
                  "validation_failed",
                  "unauthenticated",
                  "permission_denied",
                  "not_found",
                  "conflict",
                  "idempotency_key_conflict",
                  "precondition_failed",
                  "internal_error"
                ],
                "type": "string"
              },
              "details": {
                "additionalProperties": {},
                "propertyNames": {
                  "type": "string"
                },
                "type": "object"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "JournalLineRequestInput": {
        "additionalProperties": false,
        "description": "One side of one account, for a positive amount in minor units.",
        "properties": {
          "accountId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "amount": {
            "$ref": "#/components/schemas/MinorUnitsInput"
          },
          "memo": {
            "type": "string"
          },
          "side": {
            "description": "Which side of the ledger this line moves. The side carries the sign, so `amount` is always positive — a negative credit is not a debit, it is a caller that has confused the two models.",
            "enum": [
              "debit",
              "credit"
            ],
            "type": "string"
          }
        },
        "required": [
          "accountId",
          "side",
          "amount"
        ],
        "type": "object"
      },
...
      "OrgMembership": {
        "additionalProperties": false,
        "description": "One organization together with the role the caller holds in it.",
        "properties": {
          "org": {
            "$ref": "#/components/schemas/OrgSummary"
          },
          "roleCode": {
            "type": "string"
          },
          "roleId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          }
        },
        "required": [
          "org",
          "roleId",
          "roleCode"
        ],
        "type": "object"
      },
...
      "OrgMembershipList": {
        "additionalProperties": false,
        "description": "Every organization the caller is a member of, with the role held in each.",
        "properties": {
          "memberships": {
            "items": {
              "$ref": "#/components/schemas/OrgMembership"
            },
            "type": "array"
          }
        },
        "required": [
          "memberships"
        ],
        "type": "object"
      },
...
      "PostJournalRequestInput": {
        "additionalProperties": false,
        "description": "Posts one manual journal. At least two lines, and debits must equal credits exactly — there is no tolerance, because in minor units there is nothing for a tolerance to absorb.",
        "properties": {
          "date": {
            "$ref": "#/components/schemas/CalendarDateInput",
            "description": "The entry date. It must fall inside an open fiscal period — periods are never created as a side effect of posting (ROADMAP D-17), so the year has to be generated first."
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/JournalLineRequestInput"
            },
            "type": "array"
          },
          "memo": {
            "type": "string"
          }
        },
        "required": [
          "date",
          "lines"
        ],
        "type": "object"
      },
      "PostedJournal": {
        "additionalProperties": false,
        "description": "A journal as the ledger stored it, read back from the same transaction.",
        "properties": {
          "actorId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "actorType": {
            "description": "Who or what made the posting (spec §6). Taken from the session, never from the request.",
            "enum": [
              "user",
              "automation",
              "agent"
            ],
            "type": "string"
          },
          "date": {
            "$ref": "#/components/schemas/CalendarDate"
          },
          "invocationMode": {
            "anyOf": [
              {
                "enum": [
                  "interactive",
                  "scheduled"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Whether a human was present. Null means unrecorded, which is not the same as `interactive`. Set for agent callers only."
          },
          "journalId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/PostedJournalLine"
            },
            "type": "array"
          },
          "memo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "orgId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "postedAt": {
            "description": "When the journal was written. An instant, not an accounting date.",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          },
          "reversesJournalId": {
            "anyOf": [
              {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set when this journal reverses another. The link lives on the reversing journal because the original cannot be updated (ROADMAP D-02)."
          }
        },
        "required": [
          "journalId",
          "orgId",
          "date",
          "memo",
          "postedAt",
          "actorType",
          "actorId",
          "invocationMode",
          "reversesJournalId",
          "lines"
        ],
        "type": "object"
      },
...
      "PostedJournalLine": {
        "additionalProperties": false,
        "description": "One stored line. `side` is read back from which amount column is non-zero rather than echoed from the request, so this reports what the ledger holds.",
        "properties": {
          "accountId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "amount": {
            "$ref": "#/components/schemas/MinorUnits"
          },
          "lineId": {
            "type": "string"
          },
          "memo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "side": {
            "description": "Which side of the ledger this line moves. The side carries the sign, so `amount` is always positive — a negative credit is not a debit, it is a caller that has confused the two models.",
            "enum": [
              "debit",
              "credit"
            ],
            "type": "string"
          }
        },
        "required": [
          "lineId",
          "accountId",
          "side",
          "amount",
          "memo"
        ],
        "type": "object"
      },
...
      "ReverseJournalRequestInput": {
        "additionalProperties": false,
        "description": "Posts the reversal of an existing journal: a new journal with every line’s side inverted. A journal may be reversed once.",
        "properties": {
          "date": {
            "$ref": "#/components/schemas/CalendarDateInput",
            "description": "The reversal’s own entry date, which must itself fall in an open period."
          },
          "memo": {
            "type": "string"
          }
        },
        "required": [
          "date"
        ],
        "type": "object"
      },
...
      "SwitchActiveOrgRequestInput": {
        "additionalProperties": false,
        "description": "Switches the session’s active organization. The role returned is the one held in the new org, never carried across the switch.",
        "properties": {
          "orgId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          }
        },
        "required": [
          "orgId"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Open-source, AI-native, modular double-entry accounting. Every capability in the product is reachable through this API (spec §8) — the web client is a consumer of it, not a privileged path.\n\nMonetary amounts are strings of minor units (`\"1999\"` is 19.99 in a two-decimal currency), never JSON numbers. A JSON number cannot represent a cent exactly and any client that parses one has already lost the value.\n\nEvery write endpoint requires an `Idempotency-Key` header. Reusing a key with the same request replays the original outcome; reusing it with a different request is refused with `idempotency_key_conflict`.",
    "license": {
      "identifier": "AGPL-3.0-only",
      "name": "AGPL-3.0-only"
    },
    "title": "OpenBooks API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/journals": {
      "post": {
        "description": "At least two lines, debits equal to credits exactly, every account active and in this org, and the entry date inside an open fiscal period. All of it in one transaction, so a rejected posting leaves nothing behind.",
        "operationId": "postJournal",
        "parameters": [
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostJournalRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostedJournal"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Post a manual journal",
        "tags": [
          "journals"
        ]
      }
    },
    "/v1/journals/{journalId}/reverse": {
      "post": {
        "description": "Posts a new journal with every line’s side inverted and `reversesJournalId` set. The original is untouched. A journal may be reversed once; reversing the reversal re-instates the original.",
        "operationId": "reverseJournal",
        "parameters": [
          {
            "in": "path",
            "name": "journalId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReverseJournalRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostedJournal"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Reverse a posted journal",
        "tags": [
          "journals"
        ]
      }
    },
    "/v1/orgs": {
      "get": {
        "description": "The org switcher’s menu. Always the full list with the role held in each, because one login may be Owner of its own books and Read-only on a client’s.",
        "operationId": "listOrgMemberships",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgMembershipList"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Every organization the caller is a member of",
        "tags": [
          "orgs"
        ]
      },
      "post": {
        "description": "Creates an organization with the calling user as its Owner. The slug is derived from the name and collisions are disambiguated silently, so no collision is ever reported.",
        "operationId": "createOrg",
        "parameters": [
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrgRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgMembership"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Create an organization",
        "tags": [
          "orgs"
        ]
      }
    },
    "/v1/orgs/active": {
      "post": {
        "description": "Re-resolves the caller’s membership and points the session at the new org. An org the caller is not a member of and an org that does not exist both answer `not_found`.",
        "operationId": "switchActiveOrg",
        "parameters": [
          {
            "description": "Required on every write (spec §12). Send one unique value per logical request and reuse it verbatim when retrying: the same key with the same request replays the original outcome, and the same key with a different request is refused with `idempotency_key_conflict`.",
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchActiveOrgRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgMembership"
                }
              }
            },
            "description": "Default Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "summary": "Switch the session’s active organization",
        "tags": [
          "orgs"
        ]
      }
    },
...
  },
...
}